DataLongitudinal

The DataLongitudinal class handles the processing of the longitudinal data for fitting a Joint Model.

DataLongitudinal(data, formula, subject, threshold = NULL, time_grid = NULL)

Arguments

data

(data.frame)
containing the observed longitudinal data.

formula

(formula)
of the form outcome ~ time, and cannot contain any additional covariates.

subject

(string)
the name of the subject identifier variable.

threshold

(number)
cut-off value to be used to declare an observation as censored (below detection limit).

time_grid

(numeric)
grid of time points to use for providing samples of the longitudinal model fit functions. If NULL, will be taken as a sequence of 201 values from the minimum to the maximum observed times.

Details

as.list(x), as(x, "list"): Coerces x into a list of data components required for fitting a JointModel. See the vignette (TODO) for more details

as.data.frame(x)

Slots

data

(data.frame)
containing the observed longitudinal data. Note that observations that contain missing values in the required variables are removed in the slot.

formula

(formula)
of the form outcome ~ time, and cannot contain any additional covariates.

subject

(string)
the name of the subject identifier variable.

threshold

(number)
cut-off value to be used to declare an observation as censored (below detection limit).

time_grid

(numeric)
grid of time points to use for providing samples of the longitudinal model fit functions. If NULL, will be taken as a sequence of 201 values from the minimum to the maximum observed times.

See also