Construct a Variance-Covariance Matrix from Standard Deviations and Correlations

This function creates a variance-covariance matrix based on the input standard deviations and correlations. Note that the values may be altered using Matrix::nearPD() in order to ensure that it is a valid positive semi-definite matrix.

as_vcov(sd, cor)

Arguments

sd

A numeric vector containing the standard deviations of the variables.

cor

A numeric vector containing the pairwise correlations between the variables. The vector should be in the order of the upper triangular part of the matrix, excluding the diagonal elements. e.g. for a 3x3 matrix c(0.2, 0.3, 0.4) would be the values for x_12, x_13 and x_23 of the correlation matrix respectively

Value

A symmetric matrix representing the variance-covariance matrix calculated from the input standard deviations and correlations. The matrix is guaranteed to be positive semi-definite.

Examples

sd <- c(1, 2, 3)
cor <- c(0.5, 0.6, 0.7)
as_vcov(sd, cor)
#>      [,1] [,2] [,3]
#> [1,]  1.0  1.0  1.8
#> [2,]  1.0  4.0  4.2
#> [3,]  1.8  4.2  9.0