SurvivalQuantities Object & Constructor Function

Constructor function to generate a SurvivalQuantities object.

Usage

SurvivalQuantities(
  object,
  groups = NULL,
  time_grid = NULL,
  type = c("surv", "haz", "loghaz", "cumhaz")
)

Arguments

object

(JointModelSamples)
samples as drawn from a Joint Model.

groups

(character or list or NULL)
which patients to calculate the desired quantities for. See "Group Specification" for more details.

time_grid

(numeric or NULL)
vector of time points to calculate the desired quantity at. If NULL will be set to seq(0, max_survival_time, length = 201).

type

(character)
quantity to be generated. Must be one of surv, haz, loghaz, cumhaz.

Slots

quantities

(Quantities)
The sampled quantities. Should contain 1 element per element of group

groups

(list)
See argument section for details

type

(character)
See See argument section for details

time_grid

(numeric)
See argument section for details

data

(DataJoint)
The data that the Joint Model was fitted to to produce the samples/quantities

Group Specification

If groups is a character vector of subject IDs then the survival quantities will only be calculated for those specific subjects.

If groups is a list then any elements with more than 1 subject ID will be grouped together and their quantities will be calculated by taking a point-wise average. For example: groups = list("g1" = c("pt1", "pt2"), "g2" = c("pt3", "pt4")) would result in 2 groups being created whose values are the pointwise average of c("pt1", "pt2") and c("pt3", "pt4") respectively.

If groups=NULL then all subjects from original dataset will be selected

See also