jmpost: Joint Models for Predicting Overall Survival Trajectories

Implements joint models combining a non-linear mixed effects model for longitudinal measurements with a time-to-event model for the survival outcome. The acronym 'jmpost' stands for Joint Models for Predicting Overall Survival Trajectories, which refers to the application. In phase 1 studies patients are followed until progression only. Thus, reliable overall survival data and hence estimates are not available. However, we can use additional information from previous clinical trials or real-world data - we can correlate the tumour response data, that are longitudinal measurements, with the overall survival of the patients or their hazard ratios. Thereby we can predict the overall survival from our phase 1 study data and therefore make better decisions.

See also

Author

Maintainer: Craig Gower-Page craig.gower-page@roche.com

Authors:

Francois Mercier francois.mercier@roche.com

Daniel Sabanes Bove daniel.sabanes_bove@roche.com

Georgios Kazantzidis georgios.kazantzidis@roche.com

Isaac Gravestock isaac.gravestock@roche.com

Xiaoyan Fang xiaoyan.fang@roche.com

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]